experimental/extraction-framework/requirements.txt
Wand==0.6.7
google-cloud-documentai==1.2.1
img2pdf==0.4.3
flatten-json==0.1.13
pip install google-cloud-storage==2.1.0